requirements.txt
openai
python-docx
Pillow